Move language agnostic syntax comparer code to the shared layer
